requirements.txt
flask==2.3.3
psd-tools==1.9.28
pillow==10.1.0
gunicorn==21.2.0